feat(extension-api): component override map + guard/page split + imports/ui/extensible
Formal API for brand extensions to replace terminal UI components via a
components map on the workflow default export (16 canonical slots: business
pages, WelcomePage, NotFoundPage, guard fallbacks, ErrorPage, LoadingPage,
Sidebar, Header, ProminentHeader, Footer). Registry gains
registerComponent/getComponent with zIndex conflict resolution, duplicate and
unknown-key console warnings; consumption is render-time via the
useOverridableComponent hook (ExtensiblePage indirection for route tables).

Guard/page split: AuthGuard (absorbs NotSignedInWrapper + AuthenticatedRoute),
PatientGuard, DataGuard in imports/ui/guards/; overridable fallback pages
NoAuthorizationPage, NoSelectedPatientPage, NoDataPage. Default overridable
implementations moved/extracted into imports/ui/extensible/ — including
ProminentHeader, extracted from Header.jsx with its demographics prep and
medicalPolicies subscription.

Fully backwards compatible: legacy default-export keys (notFoundPage,
welcomeComponent, noPatientSelectedPage), old import paths, and the
Meteor.NoDataWrapper/NotSignedInWrapper globals all keep working behind
warn-once deprecations. Legacy welcomeComponent stays a WelcomeDialog slot,
deliberately excluded from getComponent('WelcomePage').

Docs in extensions/API.md (+ CLAUDE.md updates); unit tests via
npm run test:workflow-registry (15 tests).
