+++ imports/ui/ErrorBoundary.jsx
@@ -1,5 +1,9 @@
+// imports/ui/ErrorBoundary.jsx
+
 import React from 'react';
 import { get } from 'lodash';
+import WorkflowRegistry from '/imports/lib/WorkflowRegistry.js';
+import ErrorPage from './extensible/ErrorPage.jsx';
 
 class ErrorBoundary extends React.Component {
   constructor(props) {
@@ -19,7 +23,14 @@ class ErrorBoundary extends React.Component {
 
   render() {
     if (this.state.hasError) {
-      return this.props.fallback || <div>Something went wrong.</div>;
+      if (this.props.fallback) {
+        return this.props.fallback;
+      }
+      // Class component, so no hook here — a direct registry read is fine:
+      // errors render long after workflow registration completes.
+      const ErrorPageOverride = WorkflowRegistry.getComponent('ErrorPage');
+      const Fallback = ErrorPageOverride || ErrorPage;
+      return React.createElement(Fallback, {});
     }
 
     return this.props.children;
