+++ imports/ui/extensible/AboutPage.jsx
@@ -1,4 +1,4 @@
-// /imports/ui/pages/PrivacyPage.jsx
+// imports/ui/extensible/AboutPage.jsx
 import React from 'react';
 import { 
   Container,
@@ -10,14 +10,13 @@ import { get } from 'lodash';
 import { Meteor } from 'meteor/meteor';
 import { useTheme } from '@mui/material/styles';
 
-function PrivacyPage(props) {
+function AboutPage(props) {
   const theme = useTheme();
   
-  const title = get(Meteor, 'settings.public.businessPages.privacy.title', 'Privacy Policy');
-  const content = get(Meteor, 'settings.public.businessPages.privacy.content', 
-    `We are committed to protecting your privacy and ensuring the security of your personal information. 
-    This privacy policy outlines how we collect, use, and protect your data in accordance with HIPAA 
-    and other applicable regulations.`
+  const title = get(Meteor, 'settings.public.businessPages.about.title', 'About Us');
+  const content = get(Meteor, 'settings.public.businessPages.about.content', 
+    `Welcome to our healthcare application. We are dedicated to improving healthcare delivery 
+    through innovative technology solutions that prioritize patient care and data security.`
   );
 
   return (
@@ -49,4 +48,4 @@ function PrivacyPage(props) {
   );
 }
 
-export default PrivacyPage;
+export default AboutPage;
