+++ imports/ui/extensible/NotFoundPage.jsx
@@ -1,4 +1,4 @@
-// imports/ui/pages/NotFoundPage.jsx
+// imports/ui/extensible/NotFoundPage.jsx
 
 import React, { useState, useEffect } from 'react';
 import { get } from 'lodash';
