+++ package.json
@@ -33,6 +33,7 @@
     "postinstall": "node scripts/postinstall-dcmjs.js",
     "test:logger": "node --test imports/lib/Logger.test.mjs",
     "test:workflow-navigation": "node --test imports/lib/WorkflowNavigation.test.mjs",
+    "test:workflow-registry": "node --test imports/lib/WorkflowRegistry.test.mjs",
     "test:extensions": "node --test imports/lib/extensions/extensions.test.mjs",
     "test:reference-ranges": "node --test imports/lib/referenceRanges/qualifiers.test.mjs imports/lib/referenceRanges/resolveReferenceRange.test.mjs imports/lib/sexForClinicalUse.test.mjs imports/data/reference-ranges/blood-panel.test.mjs server/referenceRanges/seedRecords.test.mjs server/referenceRanges/mergeCandidates.test.mjs server/referenceRanges/buildContext.test.mjs",
     "test-app": "TEST_WATCH=1 meteor test --full-app --driver-package meteortesting:mocha",
